requirements.txt
# 数据
pandas>=2.0.0
numpy>=1.24.0
akshare>=1.12.0
yfinance>=0.2.28

# 配置
pyyaml>=6.0
python-dotenv>=1.0.0

# 可选：实盘券商 API 等按需安装
# vnpy  # A 股多券商
# alpaca-trade-api  # 美股 Alpaca
# 加密货币交易所
ccxt>=4.0.0,<4.1
